requirements.txt
apscheduler
duckdb
fastapi
fpdf2
httpx
numpy
pandas
python-multipart
pyyaml
uvicorn
fitparse
pytest
